Release .0 when releasing .1 if not already released
Sometimes, we release a .1 before for the first Platform
release of a given minor.
It can be because we have a CVE to fix or because we
found a major issue in the .0 before we shipped the
Platform.
In this case, we need to create the release for .0
anyway and include the major changes of it together
with the ones from the preview releases into the
.1 announcement.